feat: Unify router across run, github, and serve commands
All three commands now use create_dynamic_router(server_manager) from
package_launcher.py — the same production-grade dynamic router that
serve has always used.

Changes:
- run_servers.py: registers launched processes into server_manager.processes
  and mounts a single create_dynamic_router instead of per-server static routers
- cli.py (github_command): fixes 3-tuple unpack bug, adds full serve-parity
  setup (CORS, ServerManager, dynamic router, health/metrics, management API)
- package_launcher.py: removes create_mcp_router, create_fastapi_jsonrpc_proxy,
  start_fastapi_in_thread and stale __main__ block; launch_mcp_using_fastapi_proxy
  now returns (pkg, None, process) since callers no longer use a static router
- deprecated/router/legacy_router.py: old static router code preserved here
- CLAUDE.md: documents unified router architecture
